requirements.txt
asgiref==3.3.1
Django==3.1.3
django-taggit==1.3.0
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
